Split lint CLI action to Report and Analyze
