+++ tests/android/lint_baseline.xml
@@ -1,13 +1,12 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<issues format="6" by="lint 8.0.2" type="baseline" client="" dependencies="true" name="" variant="all" version="8.0.2">
+<issues format="5" by="lint 8.0.2" client="" dependencies="true" name="" type="baseline" variant="all" version="8.0.2">
 
     <issue
         id="LongLogTag"
         message="The logging tag can be at most 23 characters, was 39 (SomeReallyLongTagForLintToDetectAndWarn)">
         <location
             file="tests/android/src/main/java/com/grab/test/TestActivity.kt"
-            line="13"
-            column="29"/>
+            line="13"/>
     </issue>
 
 </issues>
