+++ tools/lint/src/main/java/com/grab/lint/LintResults.kt
@@ -9,7 +9,6 @@ class LintResults(
     val resultCodeFile: File,
     val lintResultsFile: File,
 ) {
-
     private fun NodeList.elements() = (0 until length).map { item(it) as Element }
 
     fun process() {
