+++ tools/lint/src/main/java/com/grab/lint/ProjectXmlCreator.kt
@@ -1,10 +1,9 @@
 package com.grab.lint
 
 import java.io.File
-import java.nio.file.Path
 
 class ProjectXmlCreator(
-    private val workingDir: Path
+    private val projectXml: File
 ) {
 
     private fun moduleXml(
@@ -35,8 +34,6 @@ class ProjectXmlCreator(
         dependencies: List<Dependency>,
         verbose: Boolean
     ): File {
-        val tempDir = workingDir.toFile()
-        val projectXml = File(tempDir, "project.xml")
         val contents = buildString {
             appendLine("<?xml version=\"1.0\" encoding=\"utf-8\"?>")
             appendLine("<project>")
